# PERF: float hash slow in py3