--- a/asv_bench/benchmarks/groupby.py
+++ b/asv_bench/benchmarks/groupby.py
@@ -515,7 +515,7 @@ class groupby_float32(object):
         tmp1 = (np.random.random(10000) * 0.1).astype(np.float32)
         tmp2 = (np.random.random(10000) * 10.0).astype(np.float32)
         tmp = np.concatenate((tmp1, tmp2))
-        arr = np.repeat(tmp, 100)
+        arr = np.repeat(tmp, 10)
         self.df = DataFrame(dict(a=arr, b=arr))
 
     def time_groupby_sum(self):