--- a/asv_bench/benchmarks/indexing.py
+++ b/asv_bench/benchmarks/indexing.py
@@ -493,7 +493,7 @@ class float_loc(object):
     goal_time = 0.2
 
     def setup(self):
-        a = np.arange(1000000)
+        a = np.arange(100000)
         self.ind = pd.Float64Index(a * 4.8000000418824129e-08)
 
     def time_float_loc(self):