@chris-b1 2016-06-15
@jreback - I've updated this.  It took a while to figure out, but the issue here isn't actually the time of the python hash function, it's that in py3 it returns a `Py_ssize_t` (was `long`), and our khash keys are 32 bit, so the truncation is causing the collisions.  I updated the asv at the top with the smaller benches - because it's a collision issue the ratios aren't both the same, but still show the issue.

```python
In [50]: a = np.arange(1000000)
    ...: ind = pd.Float64Index(a * 4.8000000418824129e-08)

In [51]: hashes = np.array([hash(x) for x in ind])

In [52]: len(hashes), len(pd.unique(hashes))
Out[52]: (1000000, 1000000)

In [53]: truncated = hashes.view('int32')[::2]

In [54]: len(truncated), len(pd.unique(truncated))
Out[54]: (1000000, 524288)
```